fixes some more issues
