--- project/src/components/movie-page/action/get-reviews.js
@@ -1,5 +1,0 @@
-import {fetchReviewList} from '../../../store/api-actions';
-
-export const getReviews = () => (dispatch, id) => {
-  dispatch(fetchReviewList(id));
-};
